Pull getPackageDirectoryPaths into a utility class
